offload SQL create table infomation from varies class (DBHandler, Model) to DAO
