+++ java/maven/factory-method/src/main/java/asia/nghiango/dbhelper/DatabaseHandler.java
@@ -9,6 +9,13 @@ import asia.nghiango.model.Model;
  * DataWriterDriver
  */
 public interface DatabaseHandler {
+    /**
+     * Creatable is to vary between database so this help
+     *
+     * @param sqlStmt 
+     */
+    public void createTable(String sqlStmt);
+
     public List<? extends Entity> getAll(String tableName, List<String> colNames);
 
     public Optional<? extends Entity> get(String tableName, List<String> colNames, int id);
